Bm dev (#42)
* readme update , vww moved to trained models

* saving y_labels for benchmark samples

* conversion performed using quantization samples

* performance samples loader

* performance samples get loaded by accessing performance samples indexes

* tflite models update based on new calibration samples

* image classification resources

* saving perf samples in bmp format

* loading perf images as bin, flattened in fortran style

* loading images with default bit orders

* image classification benchmark resources

* ic reference code update

* model substitution

* converting uint8 to int8

* tensor arena down to 100k

* calibration samples set

+++ v0.1/reference_submissions/image_classification/ic/ic_inputs.h
@@ -15,9 +15,11 @@ limitations under the License.
 #ifndef V0_1_IC_IC_INPUTS_H_
 #define V0_1_IC_IC_INPUTS_H_
 
+#include <cstdio>
+
 #include "ic/ic_model_settings.h"
 
 constexpr int kNumIcTestInputs = 1;
-extern const float g_ic_inputs[kNumIcTestInputs][kIcInputSize];
+extern const uint8_t g_ic_inputs[kNumIcTestInputs][kIcInputSize];
 
 #endif  // V0_1_IC_IC_INPUTS_H_S
